label_assistant/requirement.txt
numpy
pandas
pillow
torch
torchvision
streamlit
opencv-python-headless
huggingface-hub
# (optional) streamlit-drawable-canvas
